requirements.txt
astropy
george
lightgbm
matplotlib
numpy
pandas
requests
scikit-learn
scipy>=1.1.0
tables
tqdm
numba
statsmodels
jupyterlab
umap-learn